pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-talent-sourcing-agent"
version = "0.1.0"
description = "An agentic AI tool for automating candidate sourcing in recruiting"
authors = [
    {name = "Shashank Sangar", email = "ssangar@tesla.com"},
]
dependencies = [
    "anthropic",
    "requests",
    "python-dotenv",
    "feedparser",
    "pyyaml",
]
requires-python = ">=3.8"
readme = "README.md"

[project.scripts]
sourcing-agent = "src.sourcing_agent:main"

[tool.setuptools.packages.find]
where = ["src"]